fix unsigned int problem
Former-commit-id: aa4e0be4efe8e818b711a720b13005c272a3e9fc

+++ GeoImage/src/main/java/org/geoimage/impl/radarsat/Radarsat1Image.java
@@ -367,7 +367,7 @@ public class Radarsat1Image extends SarImageReader {
     }
 
     @Override
-    public int readPixel(int x, int y,int band) {
+    public long readPixel(int x, int y,int band) {
         int result = 0;
         long temp = 0;
         byte[] pixelByte = new byte[2];
